requirements.txt
gradio==5.14.0
azure-ai-projects==1.0.0b5
azure-identity==1.19.0
python-dotenv==1.0.1
requests==2.32.3
azure-communication-sms
azure-monitor-opentelemetry
opentelemetry-sdk
simple-salesforce